Move internal `base-button`
