+++ packages/radix-ui-themes/src/components/_internal/base-button.props.ts
@@ -1,9 +1,9 @@
-import { asChildPropDef } from '../props/as-child.prop.js';
-import { accentColorPropDef } from '../props/color.prop.js';
-import { highContrastPropDef } from '../props/high-contrast.prop.js';
-import { radiusPropDef } from '../props/radius.prop.js';
+import { asChildPropDef } from '../../props/as-child.prop.js';
+import { accentColorPropDef } from '../../props/color.prop.js';
+import { highContrastPropDef } from '../../props/high-contrast.prop.js';
+import { radiusPropDef } from '../../props/radius.prop.js';
 
-import type { PropDef } from '../props/prop-def.js';
+import type { PropDef } from '../../props/prop-def.js';
 
 const sizes = ['1', '2', '3', '4'] as const;
 const variants = ['classic', 'solid', 'soft', 'surface', 'outline', 'ghost'] as const;
