+++ packages/radix-ui-themes/src/components/_internal/base-button.tsx
@@ -3,16 +3,16 @@ import classNames from 'classnames';
 import { Slot } from '@radix-ui/react-slot';
 
 import { baseButtonPropDefs } from './base-button.props.js';
-import { Flex } from './flex.js';
-import { Spinner } from './spinner.js';
-import { VisuallyHidden } from './visually-hidden.js';
-import { extractProps } from '../helpers/extract-props.js';
-import { mapResponsiveProp, mapButtonSizeToSpinnerSize } from '../helpers/map-prop-values.js';
-import { marginPropDefs } from '../props/margin.props.js';
+import { Flex } from '../flex.js';
+import { Spinner } from '../spinner.js';
+import { VisuallyHidden } from '../visually-hidden.js';
+import { extractProps } from '../../helpers/extract-props.js';
+import { mapResponsiveProp, mapButtonSizeToSpinnerSize } from '../../helpers/map-prop-values.js';
+import { marginPropDefs } from '../../props/margin.props.js';
 
-import type { MarginProps } from '../props/margin.props.js';
-import type { ComponentPropsWithout, RemovedProps } from '../helpers/component-props.js';
-import type { GetPropDefTypes } from '../props/prop-def.js';
+import type { MarginProps } from '../../props/margin.props.js';
+import type { ComponentPropsWithout, RemovedProps } from '../../helpers/component-props.js';
+import type { GetPropDefTypes } from '../../props/prop-def.js';
 
 type BaseButtonElement = React.ElementRef<'button'>;
 type BaseButtonOwnProps = GetPropDefTypes<typeof baseButtonPropDefs>;
