+++ packages/radix-ui-themes/src/components/button.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import classNames from 'classnames';
 
-import { BaseButton } from './base-button.js';
+import { BaseButton } from './_internal/base-button.js';
 
 type ButtonElement = React.ElementRef<typeof BaseButton>;
 interface ButtonProps extends React.ComponentPropsWithoutRef<typeof BaseButton> {}
