+++ packages/radix-ui-themes/src/components/icon-button.css
@@ -1,4 +1,4 @@
-@import './base-button.css';
+@import './_internal/base-button.css';
 
 /***************************************************************************************************
  *                                                                                                 *
