+++ packages/radix-ui-themes/src/components/icon-button.props.ts
@@ -1,1 +1,1 @@
-export { baseButtonPropDefs as iconButtonPropDefs } from './base-button.props.js';
+export { baseButtonPropDefs as iconButtonPropDefs } from './_internal/base-button.props.js';
